Android: Preliminary rotation fix
