+++ src/android/app/src/main/java/io/github/borked3ds/android/viewmodel/EmulationState.kt
@@ -1,0 +1,140 @@
+// Copyright 2025 Borked3DS Emulator Project
+// Licensed under GPLv2 or any later version
+// Refer to the license.txt file included.
+
+package io.github.borked3ds.android.viewmodel
+
+import android.view.Surface
+import io.github.borked3ds.android.NativeLibrary
+import io.github.borked3ds.android.utils.Log
+
+class EmulationState(private val gamePath: String) {
+        private var state: State
+        private var surface: Surface? = null
+
+        init {
+            // Starting state is stopped.
+            state = State.STOPPED
+        }
+
+        @get:Synchronized
+        val isStopped: Boolean
+            get() = state == State.STOPPED
+
+        @get:Synchronized
+        val isPaused: Boolean
+            get() = state == State.PAUSED
+
+        @get:Synchronized
+        val isRunning: Boolean
+            get() = state == State.RUNNING
+
+        @Synchronized
+        fun stop() {
+            if (state != State.STOPPED) {
+                Log.debug("[EmulationFragment] Stopping emulation.")
+                state = State.STOPPED
+                NativeLibrary.stopEmulation()
+            } else {
+                Log.warning("[EmulationFragment] Stop called while already stopped.")
+            }
+        }
+
+        @Synchronized
+        fun pause() {
+            if (state != State.PAUSED) {
+                state = State.PAUSED
+                Log.debug("[EmulationFragment] Pausing emulation.")
+
+                // Release the surface before pausing, since emulation has to be running for that.
+                NativeLibrary.surfaceDestroyed()
+                NativeLibrary.pauseEmulation()
+            } else {
+                Log.warning("[EmulationFragment] Pause called while already paused.")
+            }
+        }
+
+        @Synchronized
+        fun unpause() {
+            if (state != State.RUNNING) {
+                state = State.RUNNING
+                Log.debug("[EmulationFragment] Unpausing emulation.")
+
+                NativeLibrary.unPauseEmulation()
+            } else {
+                Log.warning("[EmulationFragment] Unpause called while already running.")
+            }
+        }
+
+        @Synchronized
+        fun run(isActivityRecreated: Boolean) {
+            if (isActivityRecreated) {
+                if (NativeLibrary.isRunning()) {
+                    state = State.PAUSED
+                }
+            } else {
+                Log.debug("[EmulationFragment] activity resumed or fresh start")
+            }
+
+            // If the surface is set, run now. Otherwise, wait for it to get set.
+            if (surface != null) {
+                runWithValidSurface()
+            }
+        }
+
+        @Synchronized
+        fun newSurface(surface: Surface?) {
+            this.surface = surface
+            if (this.surface != null) {
+                runWithValidSurface()
+            }
+        }
+
+        @Synchronized
+        fun clearSurface() {
+            if (surface == null) {
+                Log.warning("[EmulationFragment] clearSurface called, but surface already null.")
+            } else {
+                surface = null
+                Log.debug("[EmulationFragment] Surface destroyed.")
+                when (state) {
+                    State.RUNNING -> {
+                        NativeLibrary.surfaceDestroyed()
+                        state = State.PAUSED
+                    }
+                    State.PAUSED -> {
+                        Log.warning("[EmulationFragment] Surface cleared while emulation paused.")
+                    }
+                    else -> {
+                        Log.warning("[EmulationFragment] Surface cleared while emulation stopped.")
+                    }
+                }
+            }
+        }
+
+        private fun runWithValidSurface() {
+            NativeLibrary.surfaceChanged(surface!!)
+            when (state) {
+                State.STOPPED -> {
+                    Thread({
+                        Log.debug("[EmulationFragment] Starting emulation thread.")
+                        NativeLibrary.run(gamePath)
+                    }, "NativeEmulation").start()
+                }
+                State.PAUSED -> {
+                    Log.debug("[EmulationFragment] Resuming emulation.")
+                    NativeLibrary.unPauseEmulation()
+                }
+                else -> {
+                    Log.debug("[EmulationFragment] Bug, run called while already running.")
+                }
+            }
+            state = State.RUNNING
+        }
+
+        private enum class State {
+            STOPPED,
+            RUNNING,
+            PAUSED
+        }
+}
